requirements.txt
pytest
playwright
pyyaml